+++ rer/cookieconsent/vocabularies.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
-
 from operator import itemgetter
+from plone import api
 from plone.app.vocabularies.language import AvailableContentLanguageVocabulary
-from zope.site.hooks import getSite
-from Products.CMFCore.utils import getToolByName
+# from Products.CMFCore.utils import getToolByName
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
 
@@ -12,13 +11,16 @@ class SupportedContentLanguageVocabulary(AvailableContentLanguageVocabulary):
 
     def __call__(self, context):
         items = []
-        site = getSite()
-        ltool = getToolByName(site, 'portal_languages', None)
+        ltool = api.portal.get_tool('portal_languages')
         if ltool is not None:
             items = ltool.listSupportedLanguages()
             items.sort(key=itemgetter(1))
-            items = [SimpleTerm(i[0].decode('utf-8'), i[0].decode('utf-8'), i[1].decode('utf-8')) for i in items]
+            items = [
+                SimpleTerm(
+                    i[0].decode('utf-8'),
+                    i[0].decode('utf-8'),
+                    i[1].decode('utf-8')) for i in items]
         return SimpleVocabulary(items)
 
 
-SupportedContentLanguageVocabularyFactory = SupportedContentLanguageVocabulary()
+SupportedContentLanguageVocabularyFactory = SupportedContentLanguageVocabulary()  # noqa